Refactror Helpers.cleanup_obj
We use cleaning of objects before sending them to Bugsnag. These objects
form meta data, that's why all the cleaners are under the MetaDataHash
namespace (it will be implemented later).

Splitting the method allows us to test it more easily. We have more
control of what we want to do with an object now.